requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
scikit-learn>=0.24.0
pillow>=8.3.0
tqdm>=4.62.0
wandb>=0.12.0  # for experiment tracking
albumentations>=1.0.0  # for image augmentations 